refactor: making peer pool generic
