+++ Assets/Mirage/Runtime/SocketLayer/ByteBuffer.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4a3e8b3aacf45b43bf3e29826cf13ac
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
