+++ Assets/Mirage/Runtime/SocketLayer/Connection.cs
@@ -108,7 +108,7 @@ namespace Mirage.SocketLayer
 
         IEndPoint IConnection.EndPoint => EndPoint;
 
-        internal Connection(Peer peer, IEndPoint endPoint, IDataHandler dataHandler, Config config, Time time, BufferPool bufferPool, ILogger logger, Metrics metrics)
+        internal Connection(Peer peer, IEndPoint endPoint, IDataHandler dataHandler, Config config, Time time, Pool<ByteBuffer> bufferPool, ILogger logger, Metrics metrics)
         {
             this.peer = peer;
             this.logger = logger ?? Debug.unityLogger;
