+++ Assets/Mirage/Runtime/SocketLayer/Pool.cs
@@ -4,40 +4,19 @@ using UnityEngine;
 namespace Mirage.SocketLayer
 {
     /// <summary>
-    /// Warpper around a byte[] that belongs to a <see cref="BufferPool"/>
-    /// </summary>
-    public sealed class ByteBuffer : IDisposable
-    {
-        public readonly byte[] array;
-        readonly BufferPool pool;
-
-        public ByteBuffer(int bufferSize, BufferPool pool)
-        {
-            this.pool = pool ?? throw new ArgumentNullException(nameof(pool));
-
-            array = new byte[bufferSize];
-        }
-
-        public void Release()
-        {
-            pool.Put(this);
-        }
-
-        void IDisposable.Dispose() => Release();
-    }
-
-    /// <summary>
     /// Holds a collection of <see cref="ByteBuffer"/> so they can be re-used without allocations
     /// </summary>
-    public class BufferPool
+    public class Pool<T>
     {
         const int PoolEmpty = -1;
 
         readonly int bufferSize;
         readonly int maxPoolSize;
         readonly ILogger logger;
+        readonly Func<int, Pool<T>, T> createNew;
 
-        readonly ByteBuffer[] pool;
+
+        readonly T[] pool;
         int next = -1;
         int created = 0;
 
@@ -48,29 +27,30 @@ namespace Mirage.SocketLayer
         /// <param name="startPoolSize">how many buffers to create at start</param>
         /// <param name="maxPoolSize">max number of buffers in pool</param>
         /// <param name="logger"></param>
-        public BufferPool(int bufferSize, int startPoolSize, int maxPoolSize, ILogger logger = null)
+        public Pool(Func<int, Pool<T>, T> createNew, int bufferSize, int startPoolSize, int maxPoolSize, ILogger logger = null)
         {
             if (startPoolSize > maxPoolSize) throw new ArgumentException("Start Size must be less than max size", nameof(startPoolSize));
 
             this.bufferSize = bufferSize;
             this.maxPoolSize = maxPoolSize;
             this.logger = logger ?? Debug.unityLogger;
+            this.createNew = createNew;
 
-            pool = new ByteBuffer[maxPoolSize];
+            pool = new T[maxPoolSize];
             for (int i = 0; i < startPoolSize; i++)
             {
                 Put(CreateNewBuffer());
             }
         }
 
-        private ByteBuffer CreateNewBuffer()
+        private T CreateNewBuffer()
         {
             if (created >= maxPoolSize && logger.IsLogTypeAllowed(LogType.Warning)) logger.Log(LogType.Warning, $"Buffer Max Size reached, created:{created} max:{maxPoolSize}");
             created++;
-            return new ByteBuffer(bufferSize, this);
+            return createNew.Invoke(bufferSize, this);
         }
 
-        public ByteBuffer Take()
+        public T Take()
         {
             if (next == PoolEmpty)
             {
@@ -85,7 +65,7 @@ namespace Mirage.SocketLayer
             }
         }
 
-        public void Put(ByteBuffer buffer)
+        public void Put(T buffer)
         {
             if (next < maxPoolSize - 1)
             {
